keyboard_input: gracefully handle foreground and background
With the addition of #13100 (parallel builds), people are going to try to
run Spack in the background more often.  This makes Spack handle that
situation gracefully, the way a good POSIX program should.

Specifically:

1. When `spack install` is running, we disable echo and canonical input
   so that users can type `v` to toggle build output.  We do that in a
   safe way now, so that it does not generate `SIGTTOU` in the background
   (#14682 did this).

2. We properly disable keyboard input mode when Spack is placed in the
   background, and re-enable it when Spack is in the foreground.  This
   means that if you Ctrl-Z a spack install, your terminal won't be left
   in a weird state.

3. We'll continue writing verbose output when Spack is in the background.
   If you have `stty +tostop` on, it'll end up stopping the build when
   you try to run in the background, unless you redirect output. This is
   normal behavior and it lets you easily do things like this:

       spack install -v &> log.txt &

4. Spack works fine when stopped in the background or when running in the
   background.

(2) is handled mostly with signal handlers (the way things like `vi` and
`emacs` do it) -- see the code for how that's done -- it's a bit tricky
in Python, as Python did not support blocking signals until 3.8.  It
turns out we can still make it work.